+++ src/LambdaSound/Sound/Types.hs
@@ -20,7 +20,7 @@ newtype Progress = Progress Float deriving (Show, Eq, Floating, Num, Fractional,
 newtype Percentage = Percentage Float deriving (Show, Eq, Floating, Num, Fractional, Ord, Real, RealFrac, NFData, Storable, Hashable, Enum)
 
 -- | Hz are the unit for frequencies. 440 Hz means that 440 oscillations happen per second
-newtype Hz = Hz Float deriving (Show, Eq, Ord, Num, Fractional, Floating, Enum, Generic)
+newtype Hz = Hz Float deriving (Show, Eq, Ord, Num, Fractional, Floating, Enum, Generic, Real, RealFrac)
 
 -- | Time progresses while a 'Sound' is playing and is used to create samples.
 -- It is not guaranteed that 'Time' will correspond to the real runtime of a 'Sound'